converter: `^x` on a uint8/uint16 keeps its own width, so a widened complement is not sign-extended
Go's `^x` has x's own type, so on a sub-int UNSIGNED type the complement wraps
to that width: `^uint16(5)` is 65530. C# promotes `byte`/`ushort` to `int` before
applying `~`, so bare `~x` is -6 — the same low 16 bits, but every WIDENING use
carries the sign bits, and nothing requires a cast to make that compile.

compress/flate's stored-block header is exactly that shape. Go writes

    w.writeBits(int32(^uint16(length)), 16)

and the converter emitted `(int32)(~(uint16)length)`, so `writeBits`' own
`bits |= uint64(b) << nbits` sign-extended -6 across the whole 64-bit accumulator
instead of laying down a 16-bit one's complement. Every level-0 (NoCompression)
DEFLATE stream was therefore garbage, and flate's decoder rejected its own
encoder's output: `flate: corrupt input before offset 59`. It compiled clean;
only compress/zlib's TestWriter, which round-trips every level, caught it.

The complement now truncates back to the operand's own type. Scope is exactly the
two types that need it: a SIGNED narrow type is already value-correct (C#'s `~` of
a sign-extended operand equals the sign-extended Go result — `int32(^int16(5))` is
-6 in both languages), and every type at least 32 bits wide keeps its own type
under C#'s `~`. A named type routes through its golib wrapper's operator.

Where a narrow-arithmetic cast already narrows the result the inner truncation is
redundant (`takeU8(^a)` -> `takeU8((uint8)((uint8)(~a)))`, 3 sites in one golden).
Accepted and documented: the forms are value-identical, and deciding at the unary
site whether the enclosing context widens would trade a silent-corruption hole for
readability.

Guard extends AndNotAssignNarrow, which already owns `~`-promotion on narrow
types, with the WIDENING cases plus the uint32/int16 no-op controls and the narrow
round-trip. flate's production emission is regenerated for the one affected line.

GATES: check-no-regression — exactly TWO files drifted, the guard itself and
NarrowArithmeticArg's three value-identical redundant casts; both re-baselined.
Full behavioral suite 502/502 PASS across all four phases. compress/gzip (banked
in the preceding commit, and the only validated package that links flate)
re-validated on this tree: 15/15 vs go test.

Effect on compress/zlib, which is NOT yet banked: TestWriter now passes. Its
remaining four failures are two unrelated blockers, characterized but not fixed —
three tests need the shared `../testdata` fixture tree staged (a test-host
capability: CopyFixtures rejects a parent-relative fixture by design), and
TestWriterDictIsUsed is a real preset-dictionary defect in the deflate writer
("result too large (got 104, want <= 25 bytes)").

+++ src/go-src-converted/compress/flate/huffman_bit_writer.cs
@@ -398,7 +398,7 @@ private static readonly @string writeBytesWithUnfinishedˢ = "writeBytes with un
     w.writeBits(flag, 3);
     w.flush();
     w.writeBits((int32)length, 16);
-    w.writeBits((int32)(~(uint16)length), 16);
+    w.writeBits((int32)(((uint16)(~(uint16)length))), 16);
 }
 
 [GoRecv] internal static void writeFixedHeader(this ref huffmanBitWriter w, bool isEof) {
